chore: move rate limiting to legacy folder
